Fix timestamp in YAML sample configuration
Timestamps (NN:NN:NN) needs to be surrounded by quotes in order to be
interpreted as strings.

One caveat, in the Occupancy Simulator the name of some variables are
called "offset" (e.g. step_lightsout_end_offset) and could be assumed
to be a number. However, I could not find the module occusim in the
directory appdaemon/conf/example_apps/ so I was unable to verify this.