Created a Callback Signature section of AD AD_API_REFERENCE
In learning to use AD I found myself constantly jumping back and forth
to the callback signatures for events / states / scheduler.

The current process is quite cumbersome:

1. Goto AD API
2. Open Reference
3. Go to your main area (state, scheduler, event)
4. Scroll to the listen event
5. Click to where the callback is defined in the guide.

This would have made my life much easier!

BTW - Perhaps a better way to solve this would be to have the full
documentation here in the API page. Maybe you could write the
single source of truth in the code and reference it both
in the API doc and the Guide. But I didn't want to do a bunch
of work only to have it be wrong.