constrain only the arguments
Because otherwise the type will enforce that the caller checks which variant it is, which is too verbose.

Ultimately this probably means that we shouldn't overload methods like this, but instead should have a separate interface for the non-async case where all methods are explicitly non-async, so that we can remove those annotations from the async ones and then the typer can help us to not forget "await".